+++ tests/variables/is_number.test.mjs
@@ -1,35 +1,25 @@
-import { expect, test } from '@jest/globals';
-import is_number from '../../src/variables/is_number.mjs';
+import { describe, test, expect } from '@jest/globals';
+import { is_number } from '../../src/variables.mjs';
 
-class stdClass {}
+describe('is_number', () => {
+    test('Возвращает true для числовых значений', () => {
+        expect(is_number(42)).toBe(true);
+        expect(is_number(3.14)).toBe(true);
+        expect(is_number(-0.99)).toBe(true);
+    });
 
-test('is_number', () => {
-    expect(is_number(true)).toBeFalsy();
-    expect(is_number(false)).toBeFalsy();
-    expect(is_number(0)).toBeTruthy();
-    expect(is_number(1)).toBeTruthy();
-    expect(is_number(0x539)).toBeTruthy();
-    expect(is_number(0b10100111001)).toBeTruthy();
-    expect(is_number(3.14)).toBeTruthy();
-    expect(is_number('')).toBeFalsy();
-    expect(is_number('0')).toBeFalsy();
-    expect(is_number('1')).toBeFalsy();
-    expect(is_number('0x539')).toBeFalsy();
-    expect(is_number('02471')).toBeFalsy();
-    expect(is_number('0b10100111001')).toBeFalsy();
-    expect(is_number('1337e0')).toBeFalsy();
-    expect(is_number('3.14')).toBeFalsy();
-    expect(is_number('true')).toBeFalsy();
-    expect(is_number('false')).toBeFalsy();
-    expect(is_number([])).toBeFalsy();
-    expect(is_number({})).toBeFalsy();
-    expect(is_number(new stdClass())).toBeFalsy();
-    expect(is_number(() => {})).toBeFalsy();
-    expect(is_number(null)).toBeFalsy();
-    expect(is_number(Infinity)).toBeTruthy();
-    expect(is_number(-Infinity)).toBeTruthy();
-    expect(is_number(undefined)).toBeFalsy();
-    expect(is_number(NaN)).toBeFalsy();
-    expect(is_number(stdClass)).toBeFalsy();
-    expect(is_number(Symbol())).toBeFalsy();
+    test('Возвращает false для специальных числовых значений', () => {
+        expect(is_number(NaN)).toBe(false);
+        expect(is_number(Infinity)).toBe(false);
+        expect(is_number(-Infinity)).toBe(false);
+    });
+
+    test('Возвращает false для значений не типа number', () => {
+        expect(is_number('42')).toBe(false); // строка
+        expect(is_number(true)).toBe(false); // логическое значение
+        expect(is_number(null)).toBe(false); // null
+        expect(is_number(undefined)).toBe(false); // undefined
+        expect(is_number({})).toBe(false); // объект
+        expect(is_number([])).toBe(false); // массив
+    });
 });