+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     -   [is_iterable](./docs/variables/is_iterable.md)
     -   [is_long](./docs/variables/is_long.md)
     -   [is_null](./docs/variables/is_null.md)
+    -   [is_number](./docs/variables/is_number.md)
     -   [is_object](./docs/variables/is_object.md)
 
 -   Добавлена функция [array](./docs/array/array.md) в модуль [array](./docs/array.md)